Updating code

+++ setup/bootstrap.sh
@@ -2,12 +2,12 @@
 #########################################################
 # This script is intended to be run like this:
 #
-#   curl https://.../bootstrap.sh | sudo bash
+#   curl https://mailinabox.email/setup.sh | sudo bash
 #
 #########################################################
 
 if [ -z "$TAG" ]; then
-	TAG=v0.05
+	TAG=v0.17c
 fi
 
 # Are we running as root?
@@ -18,9 +18,12 @@ fi
 
 # Clone the Mail-in-a-Box repository if it doesn't exist.
 if [ ! -d $HOME/mailinabox ]; then
-	echo Installing git . . .
-	DEBIAN_FRONTEND=noninteractive apt-get -q -q install -y git < /dev/null
-	echo
+	if [ ! -f /usr/bin/git ]; then
+		echo Installing git . . .
+		apt-get -q -q update
+		DEBIAN_FRONTEND=noninteractive apt-get -q -q install -y git < /dev/null
+		echo
+	fi
 
 	echo Downloading Mail-in-a-Box $TAG. . .
 	git clone \
